requirements.txt
amqp

annotated-types

anyio

argon2-cffi

argon2-cffi-bindings

arrow

asgiref

asttokens

attrs

billiard

binaryornot

black

boto3>=1.42.30

botocore>=1.42.30

Brotli

celery

certifi

cffi

channels

channels_redis

chardet

charset-normalizer

click

click-didyoumean

click-plugins

click-repl

colorama

coloredlogs

cookiecutter

coverage

crispy-bootstrap5

crispy-tailwind

cron_descriptor

cryptography

cssselect2

decorator

django-compressor

distro

dj-database-url

Django

django-admin-list-filter-dropdown

django-allauth

django-browser-reload

django-celery-beat

django-ckeditor-5

django-cors-headers

django-countries

django-crispy-forms

django-debug-toolbar

django-environ

django-extensions

django-filter

django-hosts

django-htmx

django_hugeicons_stroke

django-js-asset

django-mptt

django-phonenumber-field

django-rangefilter

django-ratelimit

django-redis

django-rosetta

django-storages

django-taggit

django-tailwind

django-timezone-field

django-weasyprint

django-widget-tweaks

djangorestframework

djangorestframework_simplejwt

social-auth-app-django

dkimpy

dnspython

drf-spectacular

eventlet

executing

factory_boy

Faker

flake8

flatbuffers

fonttools

gevent

greenlet

gunicorn

h11

httpcore

httpx

humanfriendly

idna

imageio

inflection

iniconfig

ipython

ipython_pygments_lexers

isort

jedi

Jinja2

jiter

jmespath

jsonschema

jsonschema-specifications

kombu

lazy_loader

llvmlite

markdown-it-py

MarkupSafe

matplotlib-inline

mccabe

mdurl

mpmath

msgpack

mypy_extensions

networkx

numba

numpy

openai

opencv-python-headless

packaging

parso

pathspec

phonenumbers

pillow

platformdirs

pluggy

polib

pooch

prompt_toolkit

protobuf

psycopg

psycopg-binary

psycopg2-binary

pure_eval

pycodestyle

pycountry

pycparser

pydantic

pydantic_core

pydyf

pyflakes

pyotp

Pygments

PyJWT

PyMatting

pyphen

pyreadline3

pytest

pytest-cov

pytest-django

python-crontab

python-dateutil

python-decouple

python-dotenv

python-slugify

PyYAML

redis

referencing

rembg

requests

beautifulsoup4

google-api-python-client

google-auth

serpapi

rich

rpds-py

s3transfer

scikit-image

scipy

setuptools

six

sniffio

sqlparse

stack-data

stripe

sympy

text-unidecode

tifffile

tinycss2

tinyhtml5

tqdm

traitlets

types-python-dateutil

typing-inspection

user-agents

typing_extensions

tzdata

uritemplate

urllib3

uvicorn

vine

wcwidth

weasyprint

webauthn

webencodings

whitenoise

zope.event

zope.interface

zopfli
